project/pyproject.toml
[project]
name = "mila-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "kuettt1"}
]
readme = "README.md"
requires-python = ">=3.9,<3.13"  # Mude para uma versão compatível
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "transformers[torch]",
    "nltk (>=3.9.1,<4.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "flask-migrate (>=4.0.0,<5.0.0)",
    "flask-jwt-extended (>=4.0.0,<5.0.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)",
    "werkzeug (>=3.0.0,<4.0.0)",
    "authlib (>=1.6.3,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"